revert to master

+++ src/mat.h
@@ -846,19 +846,19 @@ NCNN_FORCEINLINE Mat::Mat(int _w, void* _data, size_t _elemsize, Allocator* _all
 NCNN_FORCEINLINE Mat::Mat(int _w, int _h, void* _data, size_t _elemsize, Allocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(1), allocator(_allocator), dims(2), w(_w), h(_h), d(1), c(1)
 {
-    cstep = alignSize((size_t)w * h * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE Mat::Mat(int _w, int _h, int _c, void* _data, size_t _elemsize, Allocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(1), allocator(_allocator), dims(3), w(_w), h(_h), d(1), c(_c)
 {
-    cstep = alignSize((size_t)w * h * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE Mat::Mat(int _w, int _h, int _d, int _c, void* _data, size_t _elemsize, Allocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(1), allocator(_allocator), dims(4), w(_w), h(_h), d(_d), c(_c)
 {
-    cstep = alignSize((size_t)w * h * d * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * d * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE Mat::Mat(int _w, void* _data, size_t _elemsize, int _elempack, Allocator* _allocator)
@@ -870,19 +870,19 @@ NCNN_FORCEINLINE Mat::Mat(int _w, void* _data, size_t _elemsize, int _elempack, 
 NCNN_FORCEINLINE Mat::Mat(int _w, int _h, void* _data, size_t _elemsize, int _elempack, Allocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(_elempack), allocator(_allocator), dims(2), w(_w), h(_h), d(1), c(1)
 {
-    cstep = alignSize((size_t)w * h * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE Mat::Mat(int _w, int _h, int _c, void* _data, size_t _elemsize, int _elempack, Allocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(_elempack), allocator(_allocator), dims(3), w(_w), h(_h), d(1), c(_c)
 {
-    cstep = alignSize((size_t)w * h * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE Mat::Mat(int _w, int _h, int _d, int _c, void* _data, size_t _elemsize, int _elempack, Allocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(_elempack), allocator(_allocator), dims(4), w(_w), h(_h), d(_d), c(_c)
 {
-    cstep = alignSize((size_t)w * h * d * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * d * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE Mat::~Mat()
@@ -1258,38 +1258,38 @@ NCNN_FORCEINLINE const Mat Mat::channel(int _c) const
 
 NCNN_FORCEINLINE Mat Mat::depth(int z)
 {
-    Mat m(w, h, (unsigned char*)data + (size_t)w * h * z * elemsize, elemsize, elempack, allocator);
+    Mat m(w, h, (unsigned char*)data + w * h * z * elemsize, elemsize, elempack, allocator);
     m.cstep = (size_t)w * h;
     return m;
 }
 
 NCNN_FORCEINLINE const Mat Mat::depth(int z) const
 {
-    Mat m(w, h, (unsigned char*)data + (size_t)w * h * z * elemsize, elemsize, elempack, allocator);
+    Mat m(w, h, (unsigned char*)data + w * h * z * elemsize, elemsize, elempack, allocator);
     m.cstep = (size_t)w * h;
     return m;
 }
 
 NCNN_FORCEINLINE float* Mat::row(int y)
 {
-    return (float*)((unsigned char*)data + (size_t)w * y * elemsize);
+    return (float*)((unsigned char*)data + w * y * elemsize);
 }
 
 NCNN_FORCEINLINE const float* Mat::row(int y) const
 {
-    return (const float*)((unsigned char*)data + (size_t)w * y * elemsize);
+    return (const float*)((unsigned char*)data + w * y * elemsize);
 }
 
 template<typename T>
 NCNN_FORCEINLINE T* Mat::row(int y)
 {
-    return (T*)((unsigned char*)data + (size_t)w * y * elemsize);
+    return (T*)((unsigned char*)data + w * y * elemsize);
 }
 
 template<typename T>
 NCNN_FORCEINLINE const T* Mat::row(int y) const
 {
-    return (const T*)((unsigned char*)data + (size_t)w * y * elemsize);
+    return (const T*)((unsigned char*)data + w * y * elemsize);
 }
 
 NCNN_FORCEINLINE Mat Mat::channel_range(int _c, int channels)
@@ -1308,28 +1308,28 @@ NCNN_FORCEINLINE const Mat Mat::channel_range(int _c, int channels) const
 
 NCNN_FORCEINLINE Mat Mat::depth_range(int z, int depths)
 {
-    Mat m(w, h, depths, (unsigned char*)data + (size_t)w * h * z * elemsize, elemsize, elempack, allocator);
+    Mat m(w, h, depths, (unsigned char*)data + w * h * z * elemsize, elemsize, elempack, allocator);
     m.cstep = (size_t)w * h;
     return m;
 }
 
 NCNN_FORCEINLINE const Mat Mat::depth_range(int z, int depths) const
 {
-    Mat m(w, h, depths, (unsigned char*)data + (size_t)w * h * z * elemsize, elemsize, elempack, allocator);
+    Mat m(w, h, depths, (unsigned char*)data + w * h * z * elemsize, elemsize, elempack, allocator);
     m.cstep = (size_t)w * h;
     return m;
 }
 
 NCNN_FORCEINLINE Mat Mat::row_range(int y, int rows)
 {
-    Mat m(w, rows, (unsigned char*)data + (size_t)w * y * elemsize, elemsize, elempack, allocator);
+    Mat m(w, rows, (unsigned char*)data + w * y * elemsize, elemsize, elempack, allocator);
     m.cstep = (size_t)w * rows;
     return m;
 }
 
 NCNN_FORCEINLINE const Mat Mat::row_range(int y, int rows) const
 {
-    Mat m(w, rows, (unsigned char*)data + (size_t)w * y * elemsize, elemsize, elempack, allocator);
+    Mat m(w, rows, (unsigned char*)data + w * y * elemsize, elemsize, elempack, allocator);
     m.cstep = (size_t)w * rows;
     return m;
 }
@@ -1442,7 +1442,7 @@ NCNN_FORCEINLINE VkMat::VkMat(int _w, VkBufferMemory* _data, size_t _elemsize, V
 NCNN_FORCEINLINE VkMat::VkMat(int _w, int _h, VkBufferMemory* _data, size_t _elemsize, VkAllocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(1), allocator(_allocator), dims(2), w(_w), h(_h), d(1), c(1)
 {
-    cstep = alignSize((size_t)w * h * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE VkMat::VkMat(int _w, int _h, int _c, VkBufferMemory* _data, size_t _elemsize, VkAllocator* _allocator)
@@ -1466,7 +1466,7 @@ NCNN_FORCEINLINE VkMat::VkMat(int _w, VkBufferMemory* _data, size_t _elemsize, i
 NCNN_FORCEINLINE VkMat::VkMat(int _w, int _h, VkBufferMemory* _data, size_t _elemsize, int _elempack, VkAllocator* _allocator)
     : data(_data), refcount(0), elemsize(_elemsize), elempack(_elempack), allocator(_allocator), dims(2), w(_w), h(_h), d(1), c(1)
 {
-    cstep = alignSize((size_t)w * h * elemsize, 16) / elemsize;
+    cstep = alignSize(w * h * elemsize, 16) / elemsize;
 }
 
 NCNN_FORCEINLINE VkMat::VkMat(int _w, int _h, int _c, VkBufferMemory* _data, size_t _elemsize, int _elempack, VkAllocator* _allocator)
